more SP optimizations; fixed initial data sql insert; fixed group management bug
(cherry picked from commit b71db4fdec2cef862f2442ef61b0f77fab581fb6)

+++ PhpbbInDotnet.Database/Entities/PhpbbGroups.cs
@@ -27,7 +27,7 @@ namespace PhpbbInDotnet.Database.Entities
         public int GroupMessageLimit { get; set; } = 0;
         public int GroupMaxRecipients { get; set; } = 0;
         public byte GroupLegend { get; set; } = 1;
-        public int GroupUserUploadSize { get; set; } = 0;
+        public long GroupUserUploadSize { get; set; } = 0;
         public int GroupEditTime { get; set; } = 60;
     }
 }
